feat: Add v5 full backup of auth app
This commit adds a full working backup copy of authentication app. The backup is stored in a separate directory named 'backups/v5' and in 'auth-app-v5.zip'. This preserves a full working version of the authentication application implemented with sessions and cookies. This also ensures that we have a safety net in case any breaking changes happen during the development of a newer version.

- Created a new directory 'backups/v5' within the project root.
- Copied a full working version 5 where user can register, login securely with authentication.
- User can create and read a message.
- The application uses **cookies and session** for password authentication.