feat: Add callback route to authenticate requests
In this commit, we've added an authentication route for Google using Passport. The route is accessible at "/auth/google/secrets". When users visit this endpoint, they will be redirected to Google for authentication. If the authentication fails, they will be redirected to "/login". Upon successful authentication, they will be redirected to the "/secrets" page.